Feat: Implementing GetByEmail ! Incomplete

+++ SimpleNotes.Application/Services/UserService.cs
@@ -1,7 +1,9 @@
-﻿using SimpleNotes.Application.Interfaces;
+﻿using FluentResults;
+using SimpleNotes.Application.Interfaces;
 using SimpleNotes.Application.Mapping;
 using SimpleNotes.Domain;
 using SimpleNotes.Domain.Mapping;
+using SimpleNotes.Infrastructure;
 using SimpleNotes.Infrastructure.Interfaces;
 
 namespace SimpleNotes.Application.Services;
@@ -15,14 +17,47 @@ public class UserService : IUserService
         _userRepository = userRepository;
     }
 
-    public IEnumerable<UserDomain> ListUsers()
+    public Result<IEnumerable<UserDomain>> ListUsers()
     {
-        return _userRepository.ListUsers().Select(u => u.ToDomain());
+        try
+        {
+            var users = _userRepository.ListUsers().Select(x => x.ToDomain());
+
+            if (users == null)
+            {
+                return Result.Fail(new ValidationError().WithError("Id", "No user exists."));
+            }
+
+            return Result.Ok(users);
+        }
+        catch (Exception ex)
+        {
+            return Result.Fail(ex.Message);
+        }
     }
 
-    public UserDomain? GetUser(int id)
+    public Result<UserDomain?> GetUser(int id)
     {
-        return _userRepository.GetUser(id)?.ToDomain();
+        try
+        {
+            var user = _userRepository.GetUser(id)?.ToDomain();
+
+            if (user == null)
+            {
+                return Result.Fail(new ValidationError().WithError("Id", "User ID does not exist."));
+            }
+
+            return user;
+        }
+        catch (Exception ex)
+        {
+            return Result.Fail(ex.Message);
+        }
+    }
+
+    public UserDomain? GetByEmail(EmailText email)
+    {
+        return _userRepository.GetByEmail(email)?.ToDomain();
     }
 
     public IEnumerable<NoteDomain> GetUserNotes(int id)
@@ -31,31 +66,65 @@ public class UserService : IUserService
         return _userRepository.GetUserNotes(id).Select(n => n.ToDomain());
     }
 
-    public UserDomain CreateUser(UserDomain user)
+    public Result<UserDomain> CreateUser(UserDomain user)
     {
-        // Convert to Entity
-        var userEntity = user.ToEntity();
-        // Create user
-        var userCreated = _userRepository.CreateUser(userEntity);
-        // Convert to Domain
-        var userDomain = userCreated.ToDomain();
-        // return
-        return userDomain;
+        try
+        {
+            var userExists = _userRepository.GetUser(user.Id);
+
+            if (userExists == null)
+            {
+                return Result.Fail(new ValidationError().WithError("Id", "User ID does not exist"));
+            }
+
+            var userEntity = user.ToEntity();
+
+            var userCreate = _userRepository.CreateUser(userEntity);
+
+            var userDomain = userCreate.ToDomain();
+
+            return Result.Ok(userDomain);
+        }
+        catch (Exception ex)
+        {
+            return Result.Fail(ex.Message);
+        }
     }
 
-    public bool UpdateUser(UserDomain user)
+    public Result<bool> UpdateUser(UserDomain user)
     {
-        // Convert to Entity
-        var userEntity = user.ToEntity();
-        // Update
-        var userUpdated = _userRepository.UpdateUser(userEntity);
-        // Return
-        return userUpdated;
+        try
+        {
+            var userEntiy = user.ToEntity();
+
+            var userUpdate = _userRepository.UpdateUser(userEntiy);
+
+            return Result.Ok(userUpdate);
+        }
+        catch (Exception ex)
+        {
+            return Result.Fail(ex.Message);
+        }
     }
 
-    public bool DeleteUser(int id)
+    public Result<bool> DeleteUser(int id)
     {
-        // Delete user
-        return _userRepository.DeleteUser(id);
+        try
+        {
+            var userExists = _userRepository.GetUser(id);
+
+            if (userExists == null)
+            {
+                return Result.Fail(new ValidationError().WithError("Id", "User ID does not exist."));
+            }
+
+            var userDelete = _userRepository.DeleteUser(id);
+
+            return Result.Ok(userDelete);
+        }
+        catch (Exception ex)
+        {
+            return Result.Fail(ex.Message);
+        }
     }
 }
